feat(kitty): add plugin for the kitty terminal (#10994)
* A plugin for the Kitty terminal

Provides a handful of useful aliases for anyone working in a Kitty
terminal. In particular, the kssh command is handy for ensuring that
terminfo is set correctly on remote hosts when  you ssh to them.

* Guard kitty configuration

We should only load these aliases into a terminal that's identifying
itself as an xterm-kitty terminal. They're not useful in any other
situation.